Add sqliteQueries for AccountsRepository

+++ app/src/main/java/voloshyn/android/navcomponents2/screens/main/tabs/dashboard/BoxFragment.kt
@@ -48,7 +48,7 @@ class BoxFragment : Fragment(R.layout.fragment_box) {
             }
         }
 
-    private fun getBoxId(): Int {
+    private fun getBoxId(): Long {
         return args.boxId
     }
